Implemented Image, opens a window and displays it, allows user to select biomarkers.
It's coming along. Right now the program can allow a user to select the required biomarkers, but it can't calculate anything yet.

The code is also generally poor quality, I'll admit. I'm still trying to get back into the swing of things and remind myself what the best practices for programs like this really are. It's all a bit spaghetti right now, but I'll clean it up in the next commit.